Postpone the check for required paths.
The previous version was a bit too eager, and would error if paths were
required for a file that just `concat`'d other files but didn't define
any of its own.  This version waits until all the wells are loaded, and
then (if necessary) checks that every defined well has a path.