Implement the `extras` argument to load().
This gives the caller more control over how to collect information that
applies to the whole analysis, not to particular wells.  Before, all
this kind of information was supposed to go in `meta.options`.  But I
soured on this approach due to the generic/meaningless name and the
likelihood of name-clashes between different scripts.  The `extras`
argument is much more flexible, and arguably more concise (since inow
only programs that needs this kind of information will have to deal with
the corresponding return value).

Unfortunately this means I'm changing the API *again*, but I'm not at
version 1.0 yet for a reason...