Rewrite the "Basic Usage" and "Example Layouts" docs.
For the "Basic Usage" page, I wanted it to feature a more complete, and less
contrived, example.

For the "Example Layouts" page, I wanted to focus exclusively on the
layouts, i.e. not bothering to provide analysis examples.  Initially I
thought it would be good to have complete examples, but now I think that
trying to include analysis will just make things more confusing (and be
a lot of work).