fix: have [row] and [col] span all implied wells
BREAKING CHANGE:

This commit guarantees that any parameters mentioned in a [row]/[col]
block will be applied to any column/row implied in the layout.  Before,
such parameters would only be applied to columns/rows explicitly
mentioned in the layout.  The new behavior makes rows and columns more
natural to reason about: they always grow to the size of the layout.

Consider the following layout.  Previously, only wells A1 and A12
would've been included in the layout.  Now, all twelve wells in row A
are included:

    row.A.x = 1
    col.1.y = 2
    col.12.y = 3

This commit also fixes a bug where [row]/[col] didn't consider
columns/rows defined in blocks when deciding which wells to include in
the layout.   The following layout demonstrates this issue.  Previously,
this would've been an error because the row wouldn't have found any
columns to use.  Now, the row finds the columns defined by the block and
includes wells A1-A3 in the layout:

    row.A.x = 1
    block.2x2.B2.y = 2