chore: force R to install the latest version of wellmap
I've had trouble upgrading the R package, because reinstalling the
GitHub package doesn't seem to trigger the underlying python package to
be reinstalled.  This commit tries to work around this by explicitly
requesting the most recent version of the python wellmap package.  I'm
not sure if this will work, but it's worth a try.