docs: translate the "Basic usage" tutorial for R
Previously the R tutorial basically said, "read the python tutorial, but
use R syntax".  I realized that this wasn't very helpful, so I basically
copied the python tutorial and translated all the examples into R.  This
does mean that a lot of text is now duplicated, but I think it's worth
having the tutorials being more accessible to novices in either
language.