feat: simplify the *extras* argument
Previosuly, whether or not any extras were returned could change
depending on the content of the TOML file.  This made it impossible to
write robust scripts that make use of extras.  This commit makes the
*extras* argument a simple boolean instead of a list of specific keys to
return if present.  In this way, the return value is consistent and
completely independent of the input file.

Fixes #17

BREAKING CHANGE: Scipts using the *extras* argument will need to be
corrected.